Struct servum::cli::config::Config

pub struct Config {
    pub address: String,
    pub base_dir: PathBuf,
    pub list_dir: bool,
    pub port: usize,
    pub threads: usize,
    pub verbose: bool,
}

Rudimentary argument parsing and user configuration.

This approach is not very general or particularly robust, but it has the benefit of being lighter than crates such as clap.

Programs should make use of Config's default implementation, available through the Default trait by calling Config::default().

Possible configurable options include:

address: String (default: "127.0.0.1")
Server address to bind to. Default is the loopback address 127.0.0.1, i.e localhost.

base_dir: PathBuf (default: current directory)
Base directory to serve files from. Defaults to the current directory (env::current_dir)

list_dir: bool (default: true)
Whether or not to list directories. Defaults to yes.

port: usize (default: 8080)
What port to listen on. Defaults to 8080. Ports, such as port 80 (HTTP) need elevated privileges to bind to.

threads: usize (default: 4)
Number of threads to use to respond to incoming request in parallel. Value must be greater than 0. Default is 4.

verbose: bool (default: true)
Whether or not to be verbose and log stats about incoming requests. Default is true.

Example

let config = Config::default();

assert_eq!(config.address, "127.0.0.1");
assert_eq!(config.port, 8080);
// Do some more fancy stuff...

Fields
address: Stringbase_dir: PathBuflist_dir: boolport: usizethreads: usizeverbose: bool
Implementations

impl Config

pub fn new() -> Config

Create a new user configuration from environment arguments.

Read and parse environment arguments from the user and collect them into a Config struct.

If errors are encountered or the help menu is requested, the current process will be exit with code 1 (error) or 0 (help) accordingly.

fn parse_args<'a>(
args: &'a [String],
conf: &mut Config
) -> Result<bool, CliError<'a>>

Parse environment arguments and update a user Config instance

This function may return an error when it encounters an unknown argument or an invalid value.

If parsing succeeds, the function will return a boolean indicating whether the help menu has been shown (true), indicating that the program may exit normally (exit code 0), or not, meaning the program can continue.

fn help_header() -> &'static str

Return the help menu header common to all help menus.

pub fn help_long() -> String

Return the help menu in its verbose form.

pub fn help_short() -> String

Return the help menu in its short form.