Function servum::http::handler::handle_connection

pub fn handle_connection<'a>(
    req: &HTTPRequest<'_>, 
    config: Arc<Config>
) -> HTTPResponse<'a>

Handle incoming HTTP requests.

This function validates and executes incoming HTTP requests by normalizing and sanitizing request paths, respecting user config and returning a HTTPResponse with the processed file, directory listing or error.

This function solely handles HTTP GET and HEAD requests, as these are the only requests types that must be supported by simple static server.

Example

A rather extensive example involving a fair amount of boilerplate code:

use servum::cli::Config;
use std::{str, sync::Arc};

// Standard config with sane defaults
let config = Arc::new(Config::default());

// Body left out for brevity...
let buf = b"POST /contact HTTP/1.1";
let req = HTTPRequest::new(buf).unwrap();

let res = handle_connection(&req, config.clone());
let res_body = str::from_utf8(&res.body).unwrap();

// Response status
assert_eq!(res.status.to_string(), "HTTP/1.1 501 Not Implemented");

// MIME
assert_eq!(res.mime.unwrap(), "text/html");

// Response body
assert!(res_body.starts_with("<!DOCTYPE html>"));
assert!(res_body.find("<title>501</title>").is_some());
assert!(res_body.find("<h1>501</h1>").is_some());
assert!(res_body
    .find("<p>Server only supports GET and HEAD requests</p>")
    .is_some());