Struct servum::http::HTTPStatus

pub struct HTTPStatus<'a> {
    pub code: usize,
    pub msg: &'a str,
    pub comment: Option<String>,
}

A simple struct representing HTTP status responses.

This struct is used to create HTTP response messages and easily convert them to HTML documents to display them to the end-user. It also supports conversion from std::io::Result, std::io::Error and usize.

Example

let status = HTTPStatus::new(200, "OK", None);

assert_eq!(status.to_string(), "HTTP/1.1 200 OK");

Another example, using conversion from usize:

let status = HTTPStatus::from(500);

assert_eq!(status.to_string(), "HTTP/1.1 500 Internal Server Error");

A more advanced example , including conversion from std::io::Result. Note that this can be used in conjuntion with the ? operator to conveniently return HTTPStatus errors.

use std::io;

let res = io::Error::new(io::ErrorKind::PermissionDenied, "Cannot read file!");
let status = HTTPStatus::from(&res);

assert_eq!(status.to_string(), "HTTP/1.1 403 Forbidden");
assert_eq!(status.comment, Some(String::from("Cannot read file!")));

Fields
code: usizemsg: &'a strcomment: Option<String>
Implementations

impl<'a> HTTPStatus<'a>

pub fn new(code: usize, msg: &'a str, comment: Option<String>) -> Self

Create a new instance of HTTPStatus with the given code, message and optional comment

Example

let status = HTTPStatus::new(404, "Not Found", None);

assert_eq!(status.to_string(), "HTTP/1.1 404 Not Found");

pub fn to_html(&self) -> String

Create a HTML document from the status instance by calling html_doc

The status code is used as page title and lead. The optional comment is turned into a paragraph below the lead.

Example

let status = HTTPStatus::new(
    403,
    "Forbidden",
    Some(String::from("Directory traversal is forbidden!")),
);
let doc = status.to_html();

assert!(
    doc.contains("<h1>403</h1><p>Forbidden</p><p>Directory traversal is forbidden!</p>")
);